Significantly changed the structure of the reply search command
Now, the search command is in pieces. This system allows me to easily (ish) redefine the escape character. I had to add the escape character, `=`, to the problematic characters of _, %, and =. This keeps me from having to use `\` (backslash, that might have escaped the backtick depending on what this is being rendered with) as the escape character, which was having trouble being passed into SQL since python was grabbing it first. All really annoying but fixed now!
Also includes the usual amends for minor fixes